- added bottom navigation menus according to swiggy
