--- lib/ui/dashboard/dashboard_modules/cart/view/cart_controller.dart
@@ -1,13 +1,0 @@
-import 'package:taza/taza.dart';
-
-class CartController extends GetxController {
-  @override
-  void onInit() {
-    debugPrint("CartController onInit");
-    super.onInit();
-  }
-
-  void onTabSelected() {
-    debugPrint('CartController → Tab Init');
-  }
-}
