+++ lib/ui/dashboard/view/dashboard_page.dart
@@ -5,24 +5,22 @@ class DashboardPage extends GetView<DashboardController> {
 
   @override
   Widget build(BuildContext context) {
-    final style = AppTheme.of(context).tabBarStyle;
-
-    BottomNavigationBarDataModel bottomNavData =
-        controller.tabs[2].bottomNavData;
     return Scaffold(
       resizeToAvoidBottomInset: false,
       body: Stack(
         children: [
-          PageView(
-            controller: controller.pageController,
-            physics: const NeverScrollableScrollPhysics(),
-            children:
-            controller.tabs
-                .map((tab) => Builder(builder: tab.pageBuilder))
-                .toList(),
-          ),
           Obx(() {
-            return controller.currentIndex.value == 3 || controller.currentIndex.value == 1
+            return PageView(
+              controller: controller.pageController,
+              physics: const NeverScrollableScrollPhysics(),
+              children:
+              controller.tabs
+                  .map((tab) => Builder(builder: tab.pageBuilder))
+                  .toList(),
+            );
+          }),
+          Obx(() {
+            return controller.showBottomCart
                 ? Positioned.directional(
               textDirection: TextDirection.ltr,
               bottom: 0.h,
@@ -34,16 +32,7 @@ class DashboardPage extends GetView<DashboardController> {
           }),
         ],
       ),
-      floatingActionButton: FloatingActionButton(
-        elevation: 4,
-        backgroundColor: style.selectedIconColor,
-        shape: const CircleBorder(),
-        onPressed: () {
-          Get.toNamed(AppRoutes.checkoutPage);
-        },
-        child: SmartImage(path: AppImages.icShoppingBag),
-      ),
-      floatingActionButtonLocation: FloatingActionButtonLocation.centerDocked,
+
       bottomNavigationBar: Obx(() {
         return CustomBottomBar(
           selectedIndex: controller.currentIndex.value,
